services instead of data binding
