+++ src/app/app.component.html
@@ -1,3 +1,3 @@
 <app-header></app-header>
-<app-user-input (calculate)="onCalculateInvestmentResults($event)"></app-user-input>
-<app-investment-results [results]="resultsData()"></app-investment-results>
+<app-user-input ></app-user-input>
+<app-investment-results ></app-investment-results>
