+++ src/app/app.component.ts
@@ -11,42 +11,5 @@ import {type InvestmentInput } from './investment-input.model';
   imports: [HeaderComponent, UserInputComponent, InvestmentResultsComponent],
 })
 export class AppComponent {
-  resultsData=signal<{
-    year: number,
-    interest: number,
-    valueEndOfYear: number,
-    annualInvestment: number,
-    totalInterest: number,
-    totalAmountInvested: number
-  }[]|undefined>(undefined);
 
-
-  onCalculateInvestmentResults(data: InvestmentInput) {
-    console.table(data);
-    console.log(data);
-    console.log(JSON.stringify(data, null, 2));
-
-    const { expectedReturn, initialInvestment, duration, annualInvestment } =
-      data;
-    const annualData = [];
-    let investmentValue = initialInvestment;
-
-    for (let i = 0; i < duration; i++) {
-      const year = i + 1;
-      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-      investmentValue += interestEarnedInYear + annualInvestment;
-      const totalInterest =
-        investmentValue - annualInvestment * year - initialInvestment;
-      annualData.push({
-        year: year,
-        interest: interestEarnedInYear,
-        valueEndOfYear: investmentValue,
-        annualInvestment: annualInvestment,
-        totalInterest: totalInterest,
-        totalAmountInvested: initialInvestment + annualInvestment * year,
-      });
-    }
-
-    this.resultsData.set(annualData);
-  }
 }
