+++ src/app/components/investment-results/investment-results.component.html
@@ -1,4 +1,4 @@
-@if (!results()) {
+@if (!results) {
 <p class="center">Please enter some values and press "Calculate"</p>
 <!-- <p *ngIf="!results">Please enter some values and press "Calculate"</p> -->
 }
