+++ src/app/components/investment-results/investment-results.component.ts
@@ -1,5 +1,6 @@
 import { CurrencyPipe } from '@angular/common';
-import { Component, input } from '@angular/core';
+import { Component, computed, inject, input } from '@angular/core';
+import { InvestmentService } from '../../investment.service';
 @Component({
   selector: 'app-investment-results',
   standalone: true,
@@ -9,13 +10,9 @@ import { Component, input } from '@angular/core';
 })
 export class InvestmentResultsComponent {
   //results=input<...>();
- results=input<{
-  year: number,
-  interest: number,
-  valueEndOfYear: number,
-  annualInvestment: number,
-  totalInterest: number,
-  totalAmountInvested: number
-}[]>()
+  private investmentService = inject(InvestmentService);
+
+results=computed(()=>this.investmentService.resultData());
+
  
 }
