--- src/investment-results.ts
@@ -1,5 +1,0 @@
-// Use the below code as a help
-// e.g., integrate it into a service or component
-// You may need to tweak it, depending on where and how you use it
-
-
